+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oauth2/TokenGenerationService.kt
@@ -46,8 +46,10 @@ class JwtTokenGenerationService(
     internal val verifierJwsService: VerifierJwsService = DefaultVerifierJwsService(),
     /** Used to verify client attestation JWTs. */
     private val verifyJwsObject: VerifyJwsObjectFun = VerifyJwsObject(),
-    /** Used to sign DPoP (RFC 9449) access tokens, if supported by the client. */
+    @Deprecated("Use signToken instead")
     internal val jwsService: JwsService = DefaultJwsService(DefaultCryptoService(EphemeralKeyWithoutCert())),
+    /** Used to sign DPoP (RFC 9449) access tokens, if supported by the client. */
+    internal val signToken: SignJwtFun<OpenId4VciAccessToken> = SignJwt(EphemeralKeyWithoutCert(), JwsHeaderJwk()),
     /** Clock used to verify timestamps in access tokens and refresh tokens. */
     private val clock: Clock = System,
     /** Whether to issue refresh tokens, which may be used by clients to get a new access token. */
@@ -68,12 +70,9 @@ class JwtTokenGenerationService(
             TokenResponseParameters(
                 expires = 5.minutes,
                 tokenType = TOKEN_TYPE_DPOP,
-                refreshToken = if (issueRefreshToken) jwsService.createSignedJwsAddingParams(
-                    header = JwsHeader(
-                        algorithm = jwsService.algorithm,
-                        type = JwsContentTypeConstants.RT_JWT
-                    ),
-                    payload = OpenId4VciAccessToken(
+                refreshToken = if (issueRefreshToken) signToken(
+                    JwsContentTypeConstants.RT_JWT,
+                    OpenId4VciAccessToken(
                         issuer = publicContext,
                         jwtId = nonceService.provideNonce(),
                         notBefore = clock.now(),
@@ -85,17 +84,11 @@ class JwtTokenGenerationService(
                         scope = scope,
                         authorizationDetails = authorizationDetails,
                     ),
-                    serializer = OpenId4VciAccessToken.serializer(),
-                    addKeyId = false,
-                    addJsonWebKey = true,
-                    addX5c = false,
+                    OpenId4VciAccessToken.serializer(),
                 ).getOrThrow().serialize() else null,
-                accessToken = jwsService.createSignedJwsAddingParams(
-                    header = JwsHeader(
-                        algorithm = jwsService.algorithm,
-                        type = JwsContentTypeConstants.OID4VCI_AT_JWT
-                    ),
-                    payload = OpenId4VciAccessToken(
+                accessToken = signToken(
+                    JwsContentTypeConstants.OID4VCI_AT_JWT,
+                    OpenId4VciAccessToken(
                         issuer = publicContext,
                         jwtId = nonceService.provideNonce(),
                         notBefore = clock.now(),
@@ -107,10 +100,7 @@ class JwtTokenGenerationService(
                         scope = scope,
                         authorizationDetails = authorizationDetails,
                     ),
-                    serializer = OpenId4VciAccessToken.serializer(),
-                    addKeyId = false,
-                    addJsonWebKey = true,
-                    addX5c = false,
+                    OpenId4VciAccessToken.serializer(),
                 ).getOrThrow().serialize(),
                 authorizationDetails = authorizationDetails,
                 scope = scope,