+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oauth2/TokenService.kt
@@ -5,6 +5,12 @@ import at.asitplus.wallet.lib.agent.DefaultCryptoService
 import at.asitplus.wallet.lib.agent.EphemeralKeyWithoutCert
 import at.asitplus.wallet.lib.agent.KeyMaterial
 import at.asitplus.wallet.lib.jws.DefaultJwsService
+import at.asitplus.wallet.lib.jws.JwsHeaderJwk
+import at.asitplus.wallet.lib.jws.JwsHeaderJwk.invoke
+import at.asitplus.wallet.lib.jws.SignJwt
+import at.asitplus.wallet.lib.jws.SignJwt.invoke
+import at.asitplus.wallet.lib.jws.SignJwtFun
+import at.asitplus.wallet.lib.oauth2.OpenId4VciAccessToken
 import at.asitplus.wallet.lib.oidvci.DefaultNonceService
 import at.asitplus.wallet.lib.oidvci.NonceService
 
@@ -25,7 +31,7 @@ data class TokenService(
             generation = JwtTokenGenerationService(
                 nonceService = nonceService,
                 publicContext = publicContext,
-                jwsService = DefaultJwsService(DefaultCryptoService(keyMaterial)),
+                signToken = SignJwt(keyMaterial, JwsHeaderJwk()),
                 issueRefreshToken = issueRefreshTokens,
             ),
             verification = JwtTokenVerificationService(