+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/VerifiablePresentationFactory.kt
@@ -8,7 +8,6 @@ import at.asitplus.openid.dcql.DCQLClaimsQueryResult
 import at.asitplus.openid.dcql.DCQLCredentialQueryMatchingResult
 import at.asitplus.openid.third_party.at.asitplus.jsonpath.core.plus
 import at.asitplus.signum.indispensable.cosef.io.ByteStringWrapper
-import at.asitplus.signum.indispensable.josef.JwsHeader
 import at.asitplus.signum.indispensable.josef.JwsSigned
 import at.asitplus.wallet.lib.agent.SdJwtCreator.NAME_SD
 import at.asitplus.wallet.lib.data.*
@@ -16,7 +15,6 @@ import at.asitplus.wallet.lib.data.SelectiveDisclosureItem.Companion.hashDisclos
 import at.asitplus.wallet.lib.iso.*
 import at.asitplus.wallet.lib.jws.SignJwtFun
 import at.asitplus.wallet.lib.jws.JwsContentTypeConstants
-import at.asitplus.wallet.lib.jws.JwsService
 import at.asitplus.wallet.lib.jws.SdJwtSigned
 import io.github.aakira.napier.Napier
 import kotlinx.datetime.Clock
@@ -26,10 +24,9 @@ import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.json.JsonPrimitive
 
 class VerifiablePresentationFactory(
-    @Deprecated("Use signVerifiablePresentation instead")
-    private val jwsService: JwsService,
     private val identifier: String,
     private val signVerifiablePresentation: SignJwtFun<VerifiablePresentationJws>,
+    private val signKeyBinding: SignJwtFun<KeyBindingJws>,
 ) {
 
     suspend fun createVerifiablePresentationForIsoCredentials(
@@ -244,12 +241,9 @@ class VerifiablePresentationFactory(
     private suspend fun createKeyBindingJws(
         request: PresentationRequestParameters,
         issuerJwtPlusDisclosures: String,
-    ): JwsSigned<KeyBindingJws> = jwsService.createSignedJwsAddingParams(
-        header = JwsHeader(
-            type = JwsContentTypeConstants.KB_JWT,
-            algorithm = jwsService.algorithm,
-        ),
-        payload = KeyBindingJws(
+    ): JwsSigned<KeyBindingJws> = signKeyBinding(
+        JwsContentTypeConstants.KB_JWT,
+        KeyBindingJws(
             issuedAt = Clock.System.now(),
             audience = request.audience,
             challenge = request.nonce,
@@ -258,10 +252,7 @@ class VerifiablePresentationFactory(
             transactionDataHashes = if (request.transactionData?.first == PresentationRequestParameters.Flow.OID4VP) request.getTransactionDataHashes() else null,
             transactionDataHashesAlgorithm = if (request.transactionData?.first == PresentationRequestParameters.Flow.OID4VP) "sha-256" else null,
         ),
-        serializer = KeyBindingJws.serializer(),
-        addKeyId = false,
-        addJsonWebKey = false,
-        addX5c = false,
+        KeyBindingJws.serializer(),
     ).getOrElse {
         Napier.w("Could not create JWS for presentation", it)
         throw PresentationException(it)